agentcore-demo-4/requirements.txt
# Amazon Bedrock AgentCore Runtime Dependencies
# Python 3.10+ required

bedrock-agentcore>=1.0.0
boto3>=1.28.0
strands-agents>=1.0.0
mcp>=0.9.0
httpx>=0.24.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
